Phase 2: Remove redundant models and clean up tests
