--- NuTrackDemo03/Models/FoodLogEntry.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-struct FoodLogEntry: Identifiable {
-    let id = UUID()
-    let timestamp: Int64
-    let nutrition: NutritionInfo
-}
