--- NuTrackDemo03/Models/NutritionModels.swift
@@ -1,6 +1,0 @@
-struct NutritionInfo {
-    var name: String
-    var carbs: Int
-    var protein: Int
-    var fat: Int
-}
